+++ lib/connectors/DiscoverLifeAPIv2.php
@@ -293,6 +293,7 @@ class DiscoverLifeAPIv2
 
     private function delete_temp_files($file_path, $file_extension = '*')
     {
+        if(!$file_path) return;
         foreach (glob($file_path . "*." . $file_extension) as $filename) unlink($filename);
     }
