+++ lib/connectors/ObisAPI.php
@@ -453,6 +453,7 @@ class ObisAPI
 
     private function delete_temp_files($file_path, $file_extension = '*')
     {
+        if(!$file_path) return;
         foreach (glob($file_path . "*." . $file_extension) as $filename) unlink($filename);
     }
