+++ lib/connectors/TropicosAPI.php
@@ -171,6 +171,7 @@ class TropicosAPI
 
     function delete_temp_files($file_path, $file_extension = '*')
     {
+        if(!$file_path) return;
         foreach (glob($file_path . "*." . $file_extension) as $filename) unlink($filename);
     }
