+++ lib/connectors/WormsAPI.php
@@ -296,6 +296,7 @@ class WormsAPI
 
     function delete_temp_files($file_path, $file_extension = '*')
     {
+        if(!$file_path) return;
         foreach (glob($file_path . "*." . $file_extension) as $filename)
         {
            unlink($filename);
